alternancia de código do controller para o service

+++ backend/src/clients/clients.service.ts
@@ -1,6 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { isValidObjectId, Model } from 'mongoose';
 import { CreateClientDto } from './dto/create-client.dto';
 import { UpdateClientDto } from './dto/update-client.dto';
 import { Client, ClientDocument } from './entities/client.entity';
@@ -10,8 +10,22 @@ export class ClientsService {
 
   constructor(@InjectModel(Client.name) private clientModel: Model<ClientDocument>) {}
 
-  create(createClientDto: CreateClientDto) {
+  findOneByCpf(cpf: string) {
+    return this.clientModel.findOne({ cpf });
+  }
+
+  async create(createClientDto: CreateClientDto) {
     const Client = new this.clientModel(createClientDto);
+
+    const cpf = createClientDto.cpf;
+    const clientCpfExists = await this.clientModel.findOne({ cpf });
+    if (clientCpfExists) {
+      throw new HttpException(
+        'O cpf já está sendo usado',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return Client.save();
   }
 
@@ -19,15 +33,45 @@ export class ClientsService {
     return this.clientModel.find();
   }
 
-  findOne(id: string) {
+  async findOne(id: string) {
+
+    if(!isValidObjectId(id)){
+      throw new HttpException(
+        'Chave do cliente inválida',
+        HttpStatus.BAD_REQUEST,
+      );
+    }else{
+      const client = await this.clientModel.findOne({ _id: id });
+      if(!client){
+        throw new HttpException(
+          'Cliente com a chave informada não existe',
+          HttpStatus.NOT_FOUND,
+        );
+      }
+    }
+    
     return this.clientModel.findById(id);
   }
 
-  findOneByCpf(cpf: string) {
-    return this.clientModel.findOne({ cpf });
-  }
+  async update(id: string, updateClientDto: UpdateClientDto) {
+    
+    const client = await this.clientModel.findOne({ _id: id });
+    if(!client){
+      throw new HttpException(
+        'Cliente com a chave informada não existe',
+        HttpStatus.NOT_FOUND,
+      );
+    }
 
-  update(id: string, updateClientDto: UpdateClientDto) {
+    const cpf = updateClientDto.cpf;
+    const clientCpfExists = await this.clientModel.findOne({ cpf });
+    if (clientCpfExists) {
+      throw new HttpException(
+        'O cpf já está sendo usado',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return this.clientModel.findByIdAndUpdate({
       _id: id
     },{
@@ -37,7 +81,16 @@ export class ClientsService {
     });
   }
 
-  remove(id: string) {
+  async remove(id: string) {
+    
+    const client = await this.clientModel.findOne({ _id: id });
+    if(!client){
+      throw new HttpException(
+        'Cliente com a chave informada não existe',
+        HttpStatus.NOT_FOUND,
+      );
+    }
+    
     return this.clientModel.deleteOne({ _id: id }).exec();
   }
 }
